refactor: Move task into new pixi_task crate
