+++ crates/pixi_cli/Cargo.toml
@@ -77,6 +77,7 @@ pixi_pypi_spec = { workspace = true }
 pixi_record = { workspace = true }
 pixi_reporters = { workspace = true }
 pixi_spec = { workspace = true }
+pixi_task = { workspace = true }
 pixi_utils = { workspace = true, default-features = false }
 pixi_uv_conversions = { workspace = true }
 pypi_mapping = { workspace = true }
