+++ crates/pixi_cli/src/info.rs
@@ -7,11 +7,12 @@ use itertools::Itertools;
 use miette::IntoDiagnostic;
 use pixi_config;
 use pixi_consts::consts;
-use pixi_core::{WorkspaceLocator, task::TaskName};
+use pixi_core::WorkspaceLocator;
 use pixi_global::{BinDir, EnvRoot};
 use pixi_manifest::{EnvironmentName, FeatureName, SystemRequirements};
 use pixi_manifest::{FeaturesExt, HasFeaturesIter};
 use pixi_progress::await_in_progress;
+use pixi_task::TaskName;
 use rattler_conda_types::{GenericVirtualPackage, Platform};
 use rattler_networking::authentication_storage;
 use rattler_virtual_packages::{VirtualPackage, VirtualPackageOverrides};
