+++ crates/pixi_task/src/error.rs
